requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.30
pydantic>=2.9
numpy>=2.1
pandas>=2.2
scipy>=1.14
scikit-learn>=1.5
joblib>=1.4
SQLAlchemy>=2.0
pymysql>=1.1
python-dotenv>=1.0
cryptography>=42